- Breaking fix: Store `Blob` and `File` contents in binary (Latin1 8 bits) rather than     utf16 to avoid some info getting dropped when mapping to lone surrogates - Linting (ESLint): Apply script to whole repo save ignore file; override new "standard"     rule with own "object-curly-spacing"; avoid Node-11-deprecated `url.parse` - Linting (Markdown): Update per newer API - Testing (Refactoring): Use object shorthand - Testing: Don't minify test files (easier for debugging) - Testing: Check Blob contents - Testing: Check for largeValue Blob within structured cloning throwing (had conflict with user object) - Build (Rollup): Switch to terser for minification; avoid plugins where     not needed - npm: Update to Babel 7; update other devDeps; update     base64-arraybuffer-es6 deps and typeson (to ensure `userObject`     tests false for `TypesonPromise`) - npm: Bump to 1.0.0-alpha.22

+++ utils/stringArrayBuffer.js
@@ -1,5 +1,5 @@
 function arraybuffer2string (buf) {
-    return new Uint16Array(buf).reduce((s, byte) => s + String.fromCharCode(byte), '');
+    return new Uint8Array(buf).reduce((s, byte) => s + String.fromCharCode(byte), '');
 }
 
 function string2arraybuffer (str) {
@@ -49,9 +49,9 @@ function string2arraybuffer (str) {
     return buf;
     */
 
-    const array = new Uint16Array(str.length);
+    const array = new Uint8Array(str.length);
     for (let i = 0; i < str.length; i++) {
-        array[i] = str.charCodeAt(i);
+        array[i] = str.charCodeAt(i); // & 0xff;
     }
     return array.buffer;
 }
